requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
streamlit
loguru
gtts
moviepy
langchain
langchain-openai
langgraph
langdetect
PyPDF2
python-docx
transformers>=4.30.0
torch>=2.0.0
openai
